+++ addon/validator/custom-formats/int32.js
@@ -1,3 +1,5 @@
+import rangeFnFactory from './range-fn-factory'
+
 export const max = 2147483647
 export const min = -2147483648
 
@@ -6,16 +8,4 @@ export const min = -2147483648
  * @param {Any} value - value to validate
  * @returns {Boolean} whether or not value is valid
  */
-export default function (value) {
-  if (isNaN(value)) {
-    return false
-  }
-
-  const int = parseInt(value, 10)
-
-  return (
-    `${int}` === `${value}` &&
-    int >= min &&
-    int <= max
-  )
-}
+export default rangeFnFactory(min, max)
