+++ addon/validator/custom-formats/port-number.js
@@ -2,7 +2,7 @@
  * @reference https://en.wikipedia.org/wiki/Port_(computer_networking)
  */
 
-import int64 from './int64'
+import rangeFnFactory from './range-fn-factory'
 
 const min = 0
 const max = 65535
@@ -15,17 +15,4 @@ const reserved = [
  * @param {Any} value - value to validate
  * @returns {Boolean} whether or not value is valid
  */
-export default function (value) {
-  if (!int64(value)) {
-    return false
-  }
-
-  const int = parseInt(value, 10)
-
-  return (
-    `${int}` === `${value}` &&
-    int >= min &&
-    int <= max &&
-    (reserved).indexOf(int) === -1
-  )
-}
+export default rangeFnFactory(min, max, reserved)
