+++ addon/validator/custom-formats/uint16.js
@@ -1,3 +1,5 @@
+import rangeFnFactory from './range-fn-factory'
+
 export const max = 65535
 export const min = 0
 
@@ -6,16 +8,4 @@ export const min = 0
  * @param {Any} value - value to validate
  * @returns {Boolean} whether or not value is valid
  */
-export default function (value) {
-  if (isNaN(value)) {
-    return false
-  }
-
-  const int = parseInt(value, 10)
-
-  return (
-    `${int}` === `${value}` &&
-    int >= min &&
-    int <= max
-  )
-}
+export default rangeFnFactory(min, max)
