+++ addon/validator/custom-formats/vlan-id.js
@@ -2,7 +2,7 @@
  * @reference http://www.oit.ucsb.edu/committees/CNC-BEG/vlan_id.asp
  */
 
-import int64 from './int64'
+import rangeFnFactory from './range-fn-factory'
 
 export const max = 4095
 export const min = 0
@@ -16,17 +16,4 @@ export const reserved = [
  * @param {Any} value - value to validate
  * @returns {Boolean} whether or not value is valid
  */
-export default function (value) {
-  if (!int64(value)) {
-    return false
-  }
-
-  const int = parseInt(value, 10)
-
-  return (
-    `${int}` === `${value}` &&
-    int >= min &&
-    int <= max &&
-    (reserved).indexOf(int) === -1
-  )
-}
+export default rangeFnFactory(min, max, reserved)
